Move post URLs
I'm planning to re-implement community API pages as separate pages in
the UI as opposed to an endlessly growing list. The latter is a bit
problematic when opening posts beyond page 1, for people using mouse
navigation as well as people using keyboard.

Due to this, we'll need /c/[communityName]/page/[pageId] to be
available.