Load unreads and archive separately
Unfortunately we had an edge case where if you have an unread that
doesn't appear on the first page of the archive, we would not be able to
show the message in your inbox. As a result, the unread list would
incorrectly appear empty.

Since actions on the unreads and archive pages influence each other's
state, as well as the global unread counter, we sadly have to trigger 3
API calls for each read toggle interaction.